Fixed qos 0-2 in C-Broker

+++ sdk/include/dslink/col/vector.h
@@ -33,7 +33,7 @@ for (void* data = (vector)->data; n < (vector)->size; ++n, data = (char*)(vector
     /// @return 0 if the vector could be initialized successfully, otherwise -1
     int vector_init(Vector* vec, uint32_t initial_size, size_t element_size);
 
-    int vector_count(const Vector* vec);
+    uint32_t vector_count(const Vector* vec);
 
     /// Adds a value to the end of the vector, also known as push back.
     /// @param vec The vector
@@ -77,7 +77,7 @@ for (void* data = (vector)->data; n < (vector)->size; ++n, data = (char*)(vector
     /// @param data The data to find
     /// @param cmp_fn The compare function to use
     /// @return The index (>= 0) of the value if found, -1 otherwise
-    int vector_find(const Vector* vec, void* data, vector_comparison_fn_type cmp_fn);
+    long vector_find(const Vector* vec, void* data, vector_comparison_fn_type cmp_fn);
 
     /// Searches the vector for the data using the given comparison function and a binary search algorithm. The vector
     /// has to be sorted in order to work.
@@ -85,7 +85,7 @@ for (void* data = (vector)->data; n < (vector)->size; ++n, data = (char*)(vector
     /// @param data The data to find
     /// @param cmp_fn The compare function to use
     /// @return The index (>= 0) of the value if found, -1 otherwise
-    int vector_binary_search(const Vector* vec, void* data, vector_comparison_fn_type cmp_fn);
+    long vector_binary_search(const Vector* vec, void* data, vector_comparison_fn_type cmp_fn);
 
     /// Searches a range of vector for the data using the given comparison function and a binary search algorithm. 
     /// The vector has to be sorted in order to work.
@@ -95,7 +95,7 @@ for (void* data = (vector)->data; n < (vector)->size; ++n, data = (char*)(vector
     /// @param lower The lowest index of the range to search in
     /// @param upper The upper (excluded) index of the range to search in
     /// @return The index in the range [lower,upper) of the value if found, -1 otherwise
-  int vector_binary_search_range(const Vector* vec, void* data, vector_comparison_fn_type cmp_fn, 
+    long vector_binary_search_range(const Vector* vec, void* data, vector_comparison_fn_type cmp_fn, 
 				 uint32_t lower, uint32_t upper);
 
     /// Frees the internally allocated memory of the vector. Does not free the memory pointed to by the elements.
